UPDATE ::: Adding base solr engine - able to connect + query to load into a frame.  Need to expose creation of the smss/owl file directly through UI instead of hand generating UPDATE ::: Making limit/offset a long instead of an int
git-svn-id: svn://107.170.10.188:8080/Semoss/dev@21947 ef391c1c-ee15-45f7-b1d6-dde3d27a2169

Former-commit-id: 992b3f54f6680563f90d7a7a3013bea337866fbd

+++ src/prerna/ds/TinkerQueryInterpreter.java
@@ -162,8 +162,8 @@ public class TinkerQueryInterpreter extends AbstractTinkerInterpreter implements
 	 * gets the Limit/Offset and adds it to the traversal
 	 */
 	private void addLimitOffset() {
-		Integer limit = qs.getLimit();
-		Integer offset = qs.getOffset();
+		long limit = qs.getLimit();
+		long offset = qs.getOffset();
 		if (limit > 0 && offset >= 0) {
 			gt = gt.range(offset, offset + limit);
 		} else if (limit > 0) {
