+++ src/prerna/ds/querystruct/QueryStruct2.java
@@ -67,8 +67,8 @@ public class QueryStruct2 {
 //	private Hashtable<String, Set<String>> groupBy = new Hashtable<>();
 	private List<QueryStructSelector> groupBy = new ArrayList<>();
 	
-	private int limit = -1;
-	private int offset = -1;
+	private long limit = -1;
+	private long offset = -1;
 	
 	public final static int COUNT_CELLS = 1; //use this when we want to count the number of total rows*selectors in the query
 	public final static int COUNT_DISTINCT_SELECTORS = 2; //use this when we want to count the number of distinct values for a selector in a query
@@ -142,19 +142,19 @@ public class QueryStruct2 {
 		relations.put(fromConcept, compHash);
 	}
 	
-	public void setLimit(int limit) {
+	public void setLimit(long limit) {
 		this.limit = limit;
 	}
 	
-	public int getLimit() {
+	public long getLimit() {
 		return this.limit;
 	}
 	
-	public void setOffSet(int offset) {
+	public void setOffSet(long offset) {
 		this.offset = offset;
 	}
 	
-	public int getOffset() {
+	public long getOffset() {
 		return this.offset;
 	}
